WIP: Move parameter.create and .edit commands to kclCommands Fixes #7871 (eventually)
